Refactor DB from server.js
Currently most of the main app logic is all hosted in server.js.
Unfortunately, this is going to be a bit of a mess down the road, so
this is being cleaned up and the logic for each endpoint is going to be
placed in separate modules with a routing file.